cve_pipeline/requirements.txt
requests
pydantic
pydantic-settings
rich
uro
tenacity
google-generativeai
python-dotenv

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Development
black
isort
